* Renamed translation parent reference * Removed redundant error message * Move regex for timing to the top of the file * Added missing types * Added additional comments * Renamed chapter audio dialog service file
